# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyAds Dashboard API",
        "description": "API for managing advertising applications with role-based access control (Admin/Publisher)",
        "contact": {
            "name": "MyAds Support",
            "email": "admin@myads.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:5174/api",
            "description": "Frontend Proxy (Vite)"
        }
    ],
    "paths": {
        "/admin/apps": {
            "get": {
                "tags": [
                    "Admin - Apps"
                ],
                "summary": "Get all applications",
                "description": "Retrieve list of all applications from all users (Admin only)",
                "operationId": "da1f91eb83fc573eee96a03d3cc08664",
                "responses": {
                    "200": {
                        "description": "List of all applications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "My App"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "App description"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "impressions": {
                                                "type": "integer",
                                                "example": 1000
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Publisher User"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "publisher@myads.com"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/audit-logs": {
            "get": {
                "tags": [
                    "Admin - Audit"
                ],
                "summary": "Get audit logs",
                "description": "Retrieve recent audit logs (Admin only)",
                "operationId": "2368b2497dcccfb943a5b0be585710fc",
                "responses": {
                    "200": {
                        "description": "List of audit logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "action": {
                                                "type": "string",
                                                "enum": [
                                                    "created",
                                                    "updated",
                                                    "deleted",
                                                    "banned",
                                                    "unbanned"
                                                ],
                                                "example": "created"
                                            },
                                            "model": {
                                                "description": "Model type (App, User, etc.)",
                                                "type": "string",
                                                "example": "App"
                                            },
                                            "model_id": {
                                                "description": "ID of the affected model",
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "details": {
                                                "description": "Action details",
                                                "type": "string",
                                                "example": "Created new app: My App"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Publisher User"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "publisher@myads.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/stats": {
            "get": {
                "tags": [
                    "Admin - Stats"
                ],
                "summary": "Get global statistics",
                "description": "Retrieve global platform statistics (Admin only)",
                "operationId": "cbbd3c431de0e4b495c270f35af45513",
                "responses": {
                    "200": {
                        "description": "Global platform statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_users": {
                                            "description": "Total number of users",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total_publishers": {
                                            "description": "Total number of publishers",
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "total_apps": {
                                            "description": "Total number of applications",
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "active_apps": {
                                            "description": "Number of active applications",
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total_impressions": {
                                            "description": "Total impressions across all apps",
                                            "type": "string",
                                            "example": "150000"
                                        },
                                        "banned_users": {
                                            "description": "Number of banned users",
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication. Format: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}